ref: Move functions into sdk of core
