+++ packages/browser/test/index.test.ts
@@ -90,6 +90,7 @@ describe('SentryBrowser', () => {
       addBreadcrumb({ message: 'test2' });
 
       captureMessage('event');
+      popScope();
     });
   });
