+++ packages/core/src/index.ts
@@ -9,9 +9,16 @@ export {
   Severity,
   StackFrame,
   Stacktrace,
+  Scope,
   Thread,
   User,
 } from './lib/domain';
 export { DSN, DSNComponents, DSNLike, DSNProtocol } from './lib/dsn';
 export { SentryError } from './lib/error';
 export { Backend, Frontend, LogLevel, Options } from './lib/interfaces';
+export {
+  addBreadcrumb,
+  createAndBind,
+  FrontendClass,
+  setUserContext,
+} from './lib/sdk';
