+++ packages/core/test/mocks/frontend.ts
@@ -1,21 +1,14 @@
-// tslint:disable-next-line:no-submodule-imports
-import { forget } from '@sentry/utils/dist/lib/async';
 import { FrontendBase } from '../../src/lib/base';
-import { Breadcrumb, SdkInfo, User } from '../../src/lib/domain';
+import { SdkInfo } from '../../src/lib/domain';
+import { createAndBind } from '../../src/lib/sdk';
 import { TestBackend, TestOptions } from './backend';
-
-import {
-  addBreadcrumb as shimAddBreadcrumb,
-  bindClient,
-  getCurrentClient,
-  setUserContext as shimSetUserContext,
-} from '@sentry/shim';
 export {
   captureException,
   captureMessage,
   setExtraContext,
   setTagsContext,
 } from '@sentry/shim';
+export { addBreadcrumb, setUserContext } from '../../src/lib/sdk';
 
 export const TEST_SDK = {
   name: 'sentry-test',
@@ -45,17 +38,5 @@ export class TestFrontend extends FrontendBase<TestBackend, TestOptions> {
 }
 
 export function create(options: TestOptions): void {
-  if (!getCurrentClient()) {
-    const client = new TestFrontend(options);
-    forget(client.install());
-    bindClient(client);
-  }
-}
-
-export function addBreadcrumb(breadcrumb: Breadcrumb): void {
-  shimAddBreadcrumb(breadcrumb);
-}
-
-export function setUserContext(user: User): void {
-  shimSetUserContext(user);
+  createAndBind(TestFrontend, options);
 }
